Show approval time and share duration formatting
Include local approval time alongside the date and centralize compact task
duration formatting with the execution approval utilities.

Test plan: TypeScript check.
